Reorganise tests to mirror source layout
