--- llarp/crypto.h
@@ -1,6 +1,0 @@
-#ifndef LLARP_CRYPTO_H_
-#define LLARP_CRYPTO_H_
-#ifdef __cplusplus
-#include "crypto.hpp"
-#endif
-#endif
