+++ llarp/dns/dns.hpp
@@ -1,10 +1,7 @@
 #ifndef LLARP_DNS_DNS_HPP
 #define LLARP_DNS_DNS_HPP
-#include <dns/name.hpp>
-#include <dns/rr.hpp>
-#include <dns/serialize.hpp>
-#include <dns/message.hpp>
-#include <dns/server.hpp>
+
+#include <stdint.h>
 
 namespace llarp
 {
