--- llarp/exit.cpp
@@ -1,1 +1,0 @@
-#include <exit.hpp>
