--- llarp/exit.hpp
@@ -1,6 +1,0 @@
-#ifndef LLARP_EXIT_HPP
-#define LLARP_EXIT_HPP
-#include <exit/context.hpp>
-#include <exit/policy.hpp>
-#include <exit/session.hpp>
-#endif
