+++ llarp/messages/dht_immediate.hpp
@@ -1,6 +1,7 @@
 #ifndef LLARP_MESSAGES_DHT_IMMEDIATE_HPP
 #define LLARP_MESSAGES_DHT_IMMEDIATE_HPP
-#include <dht.hpp>
+
+#include <dht/message.hpp>
 #include <link_message.hpp>
 
 #include <vector>
