+++ llarp/path_types.hpp
@@ -1,8 +1,8 @@
 #ifndef LLARP_PATH_TYPES_HPP
 #define LLARP_PATH_TYPES_HPP
 
+#include <crypto.hpp>
 #include <util/aligned.hpp>
-#include <crypto.h>
 
 namespace llarp
 {
