--- llarp/service.hpp
@@ -1,14 +1,0 @@
-#ifndef LLARP_SERVICE_HPP
-#define LLARP_SERVICE_HPP
-
-#include <service/Identity.hpp>
-#include <service/Intro.hpp>
-#include <service/IntroSet.hpp>
-#include <service/config.hpp>
-#include <service/context.hpp>
-#include <service/endpoint.hpp>
-#include <service/types.hpp>
-
-#include <iostream>
-
-#endif
