+++ test/exit/test_llarp_exit_context.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include <exit.hpp>
+#include <exit/context.hpp>
 #include <router.hpp>
 
 struct ExitTest : public ::testing::Test
