--- test/test_dnsc_unit.cpp
@@ -1,6 +1,0 @@
-#include <gtest/gtest.h>
-
-#include <dnsc.hpp>
-#include <llarp.h>          // for llarp_main_init
-#include <logic.hpp>  // for threadpool/llarp::Logic
-#include <net.hpp>    // for llarp::Addr
