--- test/test_dtlslink_unit.cpp
@@ -1,9 +1,0 @@
-#include <gtest/gtest.h>
-
-struct DTLSTest : public ::testing::Test
-{
-};
-
-TEST_F(DTLSTest, TestAliceConnectToBob)
-{
-}
