+++ test/test_libabyss.cpp
