+++ test/test_llarp_dnsd.cpp
@@ -1,9 +1,10 @@
 #include <gtest/gtest.h>
 
+#include <dnsd.hpp>
+
 #include <llarp.h>          // for llarp_main_init
 #include <logic.hpp>  // for threadpool/llarp::Logic
 #include <net.hpp>    // for llarp::Addr
-#include <dnsd.hpp>
 
 unsigned int g_length = 0;
 std::string g_result  = "";
